pulled out onPaste function and other minor code refactoring
