+++ packages/plugins-image/package.json
@@ -42,6 +42,7 @@
     "prop-types": "^15.6.2",
     "react": "^16.4.1",
     "slate": ">=0.32.0",
+    "slate-react": "^0.13.2",
     "slate-schema-violations": "^0.1.15"
   }
 }
