+++ packages/plugins-image/src/renderer/index.js
@@ -1,14 +1,14 @@
 import React from "react";
 import Image from "../components/Image";
 
-const createRenderNode = (pluginOptions, pluginsWrapper) => {
-  return props => {
-    const newProps = pluginsWrapper.getProps(props);
-    switch (newProps.node.type) {
-      case "image":
-        return <Image {...newProps} options={pluginOptions} />;
-    }
-  };
+const createRenderNode = (pluginOptions, pluginsWrapper) => props => {
+  const newProps = pluginsWrapper.getProps(props);
+  switch (newProps.node.type) {
+    case "image":
+      return <Image {...newProps} options={pluginOptions} />;
+    default:
+      return undefined;
+  }
 };
 
 export default createRenderNode;
