+++ packages/plugins-image/src/static/validImageFormats.js
@@ -1,0 +1,2 @@
+const validImageFormats = ["image/gif", "image/jpeg", "image/jpg", "image/png"];
+export default validImageFormats;
